node.js/apps/WeatherApp/utils/python/socket_io_client/socket_io.py
import socketio

sio = socketio.Client()

@sio.on('connect')
def on_connect():
    print('connection established')

@sio.on('disconnect')
def on_disconnect():
    print('disconnected from server')

sio.connect('http://localhost:3000')
sio.wait()