node.js/apps/WeatherApp/utils/python/requirements.txt
###### Requirements without Version Specifiers ######
aiohttp
py4j
python-socketio
